requirements.txt
# Market Data
yfinance>=0.2.36
pandas>=1.3.0,<2.1.0
numpy>=1.21.0,<1.25.0

# Database
sqlalchemy>=1.4.0,<2.1.0

# Scheduling
schedule>=1.2.0
python-dateutil>=2.8.2

# Notifications
python-telegram-bot>=20.0
requests>=2.28.0
aiohttp>=3.8.0  # For async HTTP requests
httpx>=0.24.0  # For async HTTP requests (Strategy Composer)

# Web API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
websockets>=11.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.1

# Logging
colorlog>=6.7.0

# Visualization
plotly>=5.14.0
streamlit>=1.25.0
matplotlib>=3.5.0

# Technical Analysis
scipy>=1.10.0  # For trendline peak detection

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# V4 - Autonomous Trading
ib_insync>=0.9.86  # Interactive Brokers API wrapper

# V4.1 - RAG / Trade Memory
chromadb>=0.4.0
sentence-transformers>=2.2.0

# V4.1 - Modern Web Dashboard
nicegui>=1.4.0
